[702] added interdimensional methods.
Interdimensional methods such as .asX() and .asY() for Positionals and .asWidth() and .asHeight() for Resizables take as an argument an element with one less dimension with respect to the element which calls the method. They change one of the coordinates/dimensions of the Object with the ones of the given argument.

+++ java/src/moonaframework/design/monodimensional/Movable1DType.java
@@ -25,7 +25,7 @@ public interface Movable1DType<T extends Number> extends Positional1DType<T> {
 		
 		default @Override void setPosition(Positional1DType<Integer> pos) {
 			if (pos instanceof IntegralPositional1D actual) {
-				setX((int) actual.getX());
+				setX(actual.getX());
 			}
 			else {
 				setX(pos.getWrappedX());
@@ -50,7 +50,7 @@ public interface Movable1DType<T extends Number> extends Positional1DType<T> {
 		
 		default @Override void setPosition(Positional1DType<Float> pos) {
 			if (pos instanceof Positional1D actual) {
-				setX((float) actual.getX());
+				setX(actual.getX());
 			}
 			else {
 				setX(pos.getWrappedX());
@@ -75,7 +75,7 @@ public interface Movable1DType<T extends Number> extends Positional1DType<T> {
 		
 		default @Override void setPosition(Positional1DType<Double> pos) {
 			if (pos instanceof DoublePositional1D actual) {
-				setX((double) actual.getX());
+				setX(actual.getX());
 			}
 			else {
 				setX(pos.getWrappedX());
